Moves TestScheduleArbosUpgrade to its own file
